+++ scripts/constants.ts
@@ -1,3 +1,4 @@
 export const DIST_DIR = "dist";
 export const SRC_DIR = "src";
 export const SRC_PATH = `${SRC_DIR}/index.ts`;
+export const CDN_PATH = `${SRC_DIR}/cdn.ts`;