+++ src/core/engineLoop.ts
@@ -6,9 +6,9 @@ import {
     loadProgress,
 } from "../assets/asset";
 import type { Debug } from "../debug/debug";
+import { getDebugViewRenderer } from "../debug/debugView";
 import { SystemPhase } from "../ecs/systems/systems";
 import type { Game } from "../game/game";
-import { drawDebug } from "../gfx/draw/drawDebug";
 import { drawFrame, transformFrame } from "../gfx/draw/drawFrame";
 import { drawLoadScreen } from "../gfx/draw/drawLoadingScreen";
 import { updateViewport } from "../gfx/viewport";
@@ -111,7 +111,7 @@ export function startEngineLoop(
                 }
 
                 drawFrame();
-                if (gopt.debug !== false) drawDebug();
+                if (gopt.debug !== false) getDebugViewRenderer()?.();
 
                 for (const sys of game.systemsByEvent[SystemPhase.AfterDraw]) {
                     sys.run();